Feature

+++ feature_playlist/src/main/java/com/minhthong/playlist/domain/usecase/RemoveTrackFromPlaylistUseCase.kt
@@ -8,7 +8,7 @@ class RemoveTrackFromPlaylistUseCase @Inject constructor(
     private val repository: PlaylistRepository
 ) {
 
-    suspend operator fun invoke(playlistItemId: Int): Result<Unit> {
+    suspend operator fun invoke(playlistItemId: Long): Result<Unit> {
         return repository.removeTrackFromPlaylist(playlistItemId)
     }
 }
